requirements.txt
tensorflow>=2.10
numpy
shapely
tqdm
pillow
